ocsp: add workaround for OCSP_basic_verify() bug
Older versions of OpenSSL have a bug that it doesn't use the
certificates passed to OCSP_basic_verify() for verifying the chain. This
can be a problem when the response is signed by a certificate issued by
an intermediate CA.

      root_ca
        |
  intermediate_ca
        |-------------|
    end_entity    ocsp_signer

When the certificate hierarchy is like this, and the response contains
only ocsp_signer certificate, the following code wrongly fails.

  store = OpenSSL::X509::Store.new; store.add_cert(root_ca)
  basic_response.verify([intermediate_ca], store)

So duplicate the OCSP_BASICRESP and add the certificates to the embedded
list first.