pkey: make OpenSSL::PKey::EC::Group wrap an EC_GROUP directly
As done for EC::Point, remove ossl_ec_group struct. This contains a
breaking change. Modifications to an EC::Group returned by EC#group
no longer affects the EC object unless set to the key explicitly using
EC#group=. This is the common behavior in Ruby/OpenSSL, including other
getter methods of EC such as EC#public_key.

EC#group currently returns a EC::Group linked with the key, i.e. the
EC::Group object holds a reference to an EC_GROUP that the EC_KEY owns.
We use some ugly workaround - the ossl_ec_group struct has a flag
'dont_free' that indicates we must not free the EC_GROUP. But it is
still not possible to control OpenSSL of free'ing the EC_GROUP, so,
for example, the following code behaves strangely:

  ec = OpenSSL::PKey::EC.generate("prime256v1")
  group = ec.group
  p group.curve_name #=> "prime256v1"
  ec.group = OpenSSL::PKey::EC::Group.new("prime256v1")
  p group.curve_name #=> nil