pkey: allow specifying conversion form in EC::Point#to_bn
Currently, when we want to convert a point data into an octet string
with non-default conversion form, we have to set the desirable form to
the associated EC::Group beforehand. This is inconvenient and
counterintuitive because the conversion form is not actually related to
the EC group.

  point = ...
  point.group.point_conversion_form = :compressed
  point.to_bn

So, allow specifying the form as an optional parameter, like this:

  point = ...
  point.to_bn(:compressed)