asn1: clean up OpenSSL::ASN1::Constructive#to_der
Remove a mysterious behavior in Constructive#to_der: if the 'tagging'
attribute is set to :EXPLICIT and it is not an instance of universal
tag class classes, it "searches" the original tag from the first value
whose encoding is primitive.

	ary = [
	  OpenSSL::ASN1.Sequence([
	      OpenSSL::ASN1.OctetString("abc")
	  ])
	]
	cons = OpenSSL::ASN1::Constructive.new(ary, 1, :EXPLICIT)
	cons.to_der
	#=> "\xA1\x09\x24\x07\x30\x05\x04\x03\x61\x62\x63"
	#               ^
	#               This 4 comes from the OctetString

This is really confusing and nobody seems to be using this behavior.
Let's make it raise error instead.