bio: do not use the FILE BIO method in ossl_obj2bio()
Read everything from an IO object into a String first and use the
memory buffer BIO method just as we do for String inputs.

For MSVC builds, the FILE BIO method uses the "UPLINK" interface that
requires the application to provide OPENSSL_Applink() function. For us,
the "application" means ruby.exe, in which we can't do anything. As a
workaround, avoid using the FILE BIO method at all.

Usually private keys or X.509 certificates aren't that large and the
temporarily increased memory usage hopefully won't be an issue.

Fixes: https://github.com/ruby/openssl/issues/128