ossl_pem_passwd_cb: handle nil from the block explicitly
There is code that returns nil in the passphrase block on purpose (to
prevent OpenSSL from prompting on stdin):

	OpenSSL::PKey.read(File.read("file.pem")) { nil }

This is working just by chance because the TypeError from StringValue()
is silently ignored. Let's short circuit in that case and save raising
a needless exception, as this pattern has become too common.