test/utils: do not set ecdh_curves in start_server
An assumption in OpenSSL::TestSSL#test_get_ephemeral_key is that the
ephemeral key type is always EVP_PKEY_EC when negotiated with an ECDHE
cipher suite. This is not true if X25519 is chosen.

The test is passing because we happen to fix the group to P-256 in
start_server, but let's make it explicit.