x509name: return nil for wrong type in Name#<=>
Previously, OpenSSL::X509::Name#{cmp,<=>} would raise a TypeError if you
attempted to compare a Name object with another object of a different
type. Most Ruby classes instead return nil in this situation.

The old behavior resulted in some strange outcomes:

    >> n1 = OpenSSL::X509::Name.new

    >> 'abc' == n1
    => false

    >> n1 == 'abc'
    TypeError: wrong argument type String (expected OpenSSL/X509/NAME)

With the new behavior, cmp/<=> will return nil if the other object is
not an X509::Name instead of raising an error. This allows `==` to also
return false instead of raising an error for type mismatches.

New behavior:

    >> n1 = OpenSSL::X509::Name.new

    >> n1 == 'abc'
    => false

    >> n1 <=> 'abc'
    => nil