Remove OpenSSL::PKCS7::SignerInfo#name
This method name is misleading, because it returns the name of the
signer's issuer, not the name of the signing certificate. It is
just an alias of issuer, which is more accurate.  The "name" method
is historical, it was replaced by the "issuer" method in 2005, and
since then has been alias for backwards compatibility.

Fixes Ruby Bug 8178.