engine: revert OpenSSL::Engine.load changes for cloudhsm
Revert two commits:

 - ea49ccc82aa4 Add cloudhsm to extconf.rb
 - 33ed3ba10424 Add cloudhsm to ossl_engine.c

OpenSSL::Engine.load is a binding for ENGINE_load_*() functions which
are provided by OpenSSL itself, so-called "static engines".

Since the AWS CloudHSM engine is a dynamic engine, which is provided as
a shared library, this change is not a correct solution for the issue.

Reference: https://github.com/ruby/openssl/issues/189
Reference: https://github.com/ruby/openssl/pull/190