test(X509::Store#verify): add test case for trusted certs with same subject name
This commit attempts to reproduce https://github.com/ruby/openssl/issues/389.
Added a test suite for testing the case when we have two trusted
certificates with same subject name and we want to verify a certificate
signed by one of them. In theory, there should be no problem and this
test should run without problems, but right now, it's failing in the
`verify` of the second certificate `ee2_cert` signed by `ca1_cert_dsa`
given that the trust store is trying to verify its signature with the
first certificate `ca1_cert_rsa`, eventually resulting in an invalid
signature error.