Fix parsing EC private keys with EC parameters
The logic used with OpenSSL 1.1.1 allows for parsing private keys with
EC parameters in the PEM format. The OpenSSL 3.0 logic doesn't allow for
this and breaks loading these types of private keys.

The parsing loop here is changed to continue parsing PEM blocks also if
it succeeds to parse a PEM block, like for EC parameters. We want to
ensure that all valid PEM blocks are loaded, not only the first one.

Combined with it also trying to parse non valid PEM bits, it means we
always want to continue parsing until we hit EOF on the BIO, or if we no
longer see progress. These two checks were already implemented, so all
that's needed is to unconditionally try to grab all PEM blocks.