CI: Enable the verbose mode in the mkmf.rb by env MAKEFLAGS.
Enable the verbose option (variable `V`) in the `Makefile` generated by
`mkmf.rb` to print the compiling commands in the process of the `rake compile`
by setting the environment variable `MAKEFLAGS`[1] reserved by `make`. Because
it's helpful to see what compiler warnings are checked.

It seems that this implementation to enable the verbose mode only works in
Linux and macOS. Not Windows.

[1] https://www.gnu.org/software/make/manual/html_node/Variables_002fRecursion.html