CI: Check compiler warnings.
Add checks to make CI fail by compiler warnings in the `rake compile`.

If the `skip-warnings` (default: `false`, as an undefined variable is evaluated
as `false` in the `if` syntax) is `true` in specific matrix cases, the cases
skip the checks. If you want to skip new compiler warnings coming from external
changes such as upgraded compiler or OpenSSL versions in the specific matrix
cases, you can set the `skip-warnings: true` for the cases.