Print the Ruby and compiler info or the command itself before compiling.
The new task "debug_compiler" to print the Ruby and compiler version or,
compiler command itself for the debugging purpose. The task name is aligning
with the "debug" task that is to print the info in the Ruby OpenSSL binding for
the debugging purpose.

The compiler version info is useful when we hit the issues coming from the
newly updated compilers on the CI, and we find the cause.