ssl: support IO-like object as an underlying transport
There are use cases to establish a TLS connection on top of a non-OS
stream, such as another TLS connection or an HTTP/2 tunnel. To achieve
this today, a workaround using dummy socket pairs is necessary.

Currently, OpenSSL::SSL::SSLSocket.new requires an IO (socket) object
backed by a file descriptor. This is because we pass the file descriptor
to OpenSSL. This patch changes it to allow any Ruby object that responds
to necessary non-blocking IO methods, such as read_nonblock.

OpenSSL's TLS implementation uses an IO abstraction layer called BIO to
interact with the underlying socket. By passing the file descriptor to
SSL_set_fd(), a BIO with the BIO_s_socket() BIO_METHOD is implicitly
created. We can set up our own BIO and let OpenSSL use it instead. The
previous patch added such a BIO_METHOD implementation.

For performance reason, this patch continues to use the socket BIO if
the user passes a real IO object, so this should not change the behavior
of existing programs in any way.