ssl: allow failure on test_connect_certificate_verify_failed_exception_message
[ This patch only applies to the 3.0 and 3.1 branch. ]

It is a test case for SSLSocket generating an informative error message
on a certificate verification failure. A change in OpenSSL 3.1 broke it
and a generic error message is currently generated.

This is fixed in the 3.2 branch by commit 5113777e8271, but I decided
not to backport the commit to the 3.0 branch because the diff doesn't
apply cleanly.