CI: List OpenSSL providers by openssl CLI in FIPS cases.
Add the step to list OpenSSL providers available on OpenSSL 3.0 or later
versions to check if the base and fips providers are loaded in the FIPS cases.

We have a logic to print the OpenSSL providers in the Rake test and test_fips
tasks calling the debug task.
https://github.com/ruby/openssl/blob/16aa2b2f0c10d9d2b6ccf09628d248ba9d2f3f4e/Rakefile#L72

However, if the fips provider is not loaded, the Ruby OpenSSL aborts before
printing the OpenSSL providers in the Rake debug task.
https://github.com/ruby/openssl/actions/runs/10077703798/job/27860837398#step:13:35

This commit enables CI to print the loaded OpenSSL providers in such case.

Added the `'-Wl,-rpath,$(LIBRPATH)'` to call the openssl cli without setting
`LD_LIBRARY_PATH`.
See https://github.com/ruby/openssl/blob/master/CONTRIBUTING.md#with-different-versions-of-openssl for details.